Add files via upload
This algorithm is for electrical engineers to calculate the nodal admittance matrix (Ybus) for three phase systems .
In power engineering, Ybus or Y Matrix is an N x N matrix describing a linear power system with N buses. It represents the nodal admittance of the buses in a power system.